senses when a clog has occurred, fully opens and closes valve system

+++ software/ADPL_electron/Bucket.cpp
@@ -24,7 +24,7 @@ void Bucket::updateFlow(){
     highestFlow = (unsigned long) _VOLUME*3600*(1/_HIGHESTFLOW);
 
     timeRead = millis();
-    tipTime = (unsigned long)(timeRead-lastTime);
+    tipTime = timeRead-lastTime;
     lastTime = timeRead;
     tip = false;
 }
